requirements.txt
django
django_heroku
newsapi-python
dj-database-url
psycopg2-binary
django-bmemcached
gunicorn